feat: Add mobile command for file deletion
