+++ lib/commands/actions.js
@@ -1,6 +1,5 @@
 import androidHelpers from '../android-helpers';
-import _ from 'lodash';
-import { fs, util, zip, tempDir} from 'appium-support';
+import { fs, util, tempDir} from 'appium-support';
 import path from 'path';
 import log from '../logger';
 import B from 'bluebird';
@@ -9,10 +8,6 @@ import { exec } from 'teen_process';
 
 const swipeStepsPerSec = 28;
 const dragStepsPerSec = 40;
-const CONTAINER_PATH_MARKER = '@';
-// https://regex101.com/r/PLdB0G/2
-const CONTAINER_PATH_PATTERN = new RegExp(`^${CONTAINER_PATH_MARKER}([^/]+)/(.+)`);
-const ANDROID_MEDIA_RESCAN_INTENT = 'android.intent.action.MEDIA_SCANNER_SCAN_FILE';
 
 let commands = {}, helpers = {}, extensions = {};
 
@@ -149,112 +144,6 @@ commands.openNotifications = async function openNotifications () {
 
 commands.setLocation = async function setLocation (latitude, longitude) {
   return await this.adb.sendTelnetCommand(`geo fix ${longitude} ${latitude}`);
-};
-
-function parseContainerPath (remotePath) {
-  const match = CONTAINER_PATH_PATTERN.exec(remotePath);
-  if (!match) {
-    log.errorAndThrow(`It is expected that package identifier is separated from the relative path with a single slash. ` +
-                      `'${remotePath}' is given instead`);
-  }
-  return [match[1], path.posix.resolve(`/data/data/${match[1]}`, match[2])];
-}
-
-commands.pullFile = async function pullFile (remotePath) {
-  if (remotePath.endsWith('/')) {
-    log.errorAndThrow(`It is expected that remote path points to a file and not to a folder. ` +
-                      `'${remotePath}' is given instead`);
-  }
-  let tmpDestination = null;
-  if (remotePath.startsWith(CONTAINER_PATH_MARKER)) {
-    const [packageId, pathInContainer] = parseContainerPath(remotePath);
-    log.info(`Parsed package identifier '${packageId}' from '${remotePath}'. Will get the data from '${pathInContainer}'`);
-    tmpDestination = `/data/local/tmp/${path.posix.basename(pathInContainer)}`;
-    try {
-      await this.adb.shell(['run-as', packageId, `chmod 777 '${pathInContainer.replace(/'/g, '\\\'')}'`]);
-      await this.adb.shell(['cp', '-f', pathInContainer, tmpDestination]);
-    } catch (e) {
-      log.errorAndThrow(`Cannot access the container of '${packageId}' application. ` +
-                        `Is the application installed and has 'debuggable' build option set to true? ` +
-                        `Original error: ${e.message}`);
-    }
-  }
-  const localFile = await tempDir.path({prefix: 'appium', suffix: '.tmp'});
-  try {
-    await this.adb.pull(_.isString(tmpDestination) ? tmpDestination : remotePath, localFile);
-    return (await util.toInMemoryBase64(localFile)).toString();
-  } finally {
-    if (await fs.exists(localFile)) {
-      await fs.unlink(localFile);
-    }
-    if (_.isString(tmpDestination)) {
-      await this.adb.shell(['rm', '-f', tmpDestination]);
-    }
-  }
-};
-
-commands.pushFile = async function pushFile (remotePath, base64Data) {
-  if (remotePath.endsWith('/')) {
-    log.errorAndThrow(`It is expected that remote path points to a file and not to a folder. ` +
-                      `'${remotePath}' is given instead`);
-  }
-  const localFile = await tempDir.path({prefix: 'appium', suffix: '.tmp'});
-  if (_.isArray(base64Data)) {
-    // some clients (ahem) java, send a byte array encoding utf8 characters
-    // instead of a string, which would be infinitely better!
-    base64Data = Buffer.from(base64Data).toString('utf8');
-  }
-  const content = Buffer.from(base64Data, 'base64');
-  let tmpDestination = null;
-  try {
-    await fs.writeFile(localFile, content.toString('binary'), 'binary');
-    if (remotePath.startsWith(CONTAINER_PATH_MARKER)) {
-      const [packageId, pathInContainer] = parseContainerPath(remotePath);
-      log.info(`Parsed package identifier '${packageId}' from '${remotePath}'. Will put the data into '${pathInContainer}'`);
-      tmpDestination = `/data/local/tmp/${path.posix.basename(pathInContainer)}`;
-      try {
-        await this.adb.shell(
-          ['run-as', packageId, `mkdir -p '${path.posix.dirname(pathInContainer).replace(/'/g, '\\\'')}'`]
-        );
-        await this.adb.shell(['run-as', packageId, `touch '${pathInContainer.replace(/'/g, '\\\'')}'`]);
-        await this.adb.shell(['run-as', packageId, `chmod 777 '${pathInContainer.replace(/'/g, '\\\'')}'`]);
-        await this.adb.push(localFile, tmpDestination);
-        await this.adb.shell(['cp', '-f', tmpDestination, pathInContainer]);
-      } catch (e) {
-        log.errorAndThrow(`Cannot access the container of '${packageId}' application. ` +
-                          `Is the application installed and has 'debuggable' build option set to true? ` +
-                          `Original error: ${e.message}`);
-      }
-    } else {
-      // adb push creates folders and overwrites existing files.
-      await this.adb.push(localFile, remotePath);
-
-      // if we have pushed a file, it might be a media file, so ensure that
-      // apps know about it
-      log.info('After pushing media file, broadcasting media scan intent');
-      try {
-        await this.adb.shell(['am', 'broadcast', '-a',
-          ANDROID_MEDIA_RESCAN_INTENT, '-d', `file://${remotePath}`]);
-      } catch (e) {
-        log.warn(`Got error broadcasting media scan intent: ${e.message}; ignoring`);
-      }
-    }
-  } finally {
-    if (await fs.exists(localFile)) {
-      await fs.unlink(localFile);
-    }
-    if (_.isString(tmpDestination)) {
-      await this.adb.shell(['rm', '-f', tmpDestination]);
-    }
-  }
-};
-
-commands.pullFolder = async function pullFolder (remotePath) {
-  let localFolder = await tempDir.path({prefix: 'appium'});
-  await this.adb.pull(remotePath, localFolder);
-  return (await zip.toInMemoryZip(localFolder, {
-    encodeToBase64: true,
-  })).toString();
 };
 
 commands.fingerprint = async function fingerprint (fingerprintId) {
