+++ lib/commands/execute.js
@@ -46,6 +46,8 @@ extensions.executeMobile = async function executeMobile (mobileCommand, opts = {
     getNotifications: 'mobileGetNotifications',
 
     listSms: 'mobileListSms',
+
+    deleteFile: 'mobileDeleteFile',
   };
 
   if (!_.has(mobileCommandsMapping, mobileCommand)) {
