+++ lib/commands/index.js
@@ -14,6 +14,7 @@ import performanceCmds from './performance';
 import executeCmds from './execute';
 import shellCmds from './shell';
 import emuConsoleCmds from './emu-console';
+import fileActionsCmds from './file-actions';
 import appManagementCmds from './app-management';
 import logCmds from './log';
 
@@ -38,6 +39,7 @@ Object.assign(
   shellCmds,
   emuConsoleCmds,
   appManagementCmds,
+  fileActionsCmds,
   logCmds,
   // add other command types here
 );
